mo'controllers, mo'comments
In a desire to make this application not confusing as hell, both to the
team and to our posterity, I have added a ton of comments and made more
controllers to make their functions more clear. Hopefully I have used
intuitive naming conventions that will help users in the future.